feat: add test-utils with groth16 `dummy_vk`
