+++ src/gadgets/groth16.rs
@@ -508,6 +508,8 @@ mod tests {
 
     // Helper to reduce duplication across bitflip tests for A, B, and C
     fn run_false_bitflip_test(seed: u64, mutate: impl FnOnce(&mut Groth16VerifyInput)) {
+        use crate::test_utils::DummyCircuit;
+
         let k = 6;
         let mut rng = ChaCha20Rng::seed_from_u64(seed);
         let circuit = DummyCircuit::<ark_bn254::Fr> {
@@ -537,83 +539,18 @@ mod tests {
         assert!(!out.output_value);
     }
 
-    #[derive(Copy, Clone)]
-    struct DummyCircuit<F: ark_ff::PrimeField> {
-        pub a: Option<F>,
-        pub b: Option<F>,
-        pub num_variables: usize,
-        pub num_constraints: usize,
-    }
-
-    impl<F: ark_ff::PrimeField> ConstraintSynthesizer<F> for DummyCircuit<F> {
-        fn generate_constraints(self, cs: ConstraintSystemRef<F>) -> Result<(), SynthesisError> {
-            let a = cs.new_witness_variable(|| self.a.ok_or(SynthesisError::AssignmentMissing))?;
-            let b = cs.new_witness_variable(|| self.b.ok_or(SynthesisError::AssignmentMissing))?;
-            let c = cs.new_input_variable(|| {
-                let a = self.a.ok_or(SynthesisError::AssignmentMissing)?;
-                let b = self.b.ok_or(SynthesisError::AssignmentMissing)?;
-                Ok(a * b)
-            })?;
-
-            for _ in 0..(self.num_variables - 3) {
-                let _ =
-                    cs.new_witness_variable(|| self.a.ok_or(SynthesisError::AssignmentMissing))?;
-            }
-
-            for _ in 0..self.num_constraints - 1 {
-                cs.enforce_constraint(lc!() + a, lc!() + b, lc!() + c)?;
-            }
-
-            cs.enforce_constraint(lc!(), lc!(), lc!())?;
-            Ok(())
-        }
-    }
-
-    // Circuit with no public inputs - only private witnesses
-    #[derive(Copy, Clone)]
-    struct DummyCircuitNoPublicInputs<F: ark_ff::PrimeField> {
-        pub a: Option<F>,
-        pub b: Option<F>,
-        pub num_variables: usize,
-        pub num_constraints: usize,
-    }
-
-    impl<F: ark_ff::PrimeField> ConstraintSynthesizer<F> for DummyCircuitNoPublicInputs<F> {
-        fn generate_constraints(self, cs: ConstraintSystemRef<F>) -> Result<(), SynthesisError> {
-            let a = cs.new_witness_variable(|| self.a.ok_or(SynthesisError::AssignmentMissing))?;
-            let b = cs.new_witness_variable(|| self.b.ok_or(SynthesisError::AssignmentMissing))?;
-            let c = cs.new_witness_variable(|| {
-                let a = self.a.ok_or(SynthesisError::AssignmentMissing)?;
-                let b = self.b.ok_or(SynthesisError::AssignmentMissing)?;
-                Ok(a * b)
-            })?;
-
-            for _ in 0..(self.num_variables - 3) {
-                let _ =
-                    cs.new_witness_variable(|| self.a.ok_or(SynthesisError::AssignmentMissing))?;
-            }
-
-            for _ in 0..self.num_constraints - 1 {
-                cs.enforce_constraint(lc!() + a, lc!() + b, lc!() + c)?;
-            }
-
-            cs.enforce_constraint(lc!(), lc!(), lc!())?;
-            Ok(())
-        }
-    }
-
     #[test]
     fn test_groth16_verify_true() {
-        let k = 6;
-        let mut rng = ChaCha20Rng::seed_from_u64(12345);
+        use crate::test_utils::{DummyCircuit, dummy_vk_with_public_inputs};
+
+        let (pk, vk, a, b, c_val) = dummy_vk_with_public_inputs();
         let circuit = DummyCircuit::<ark_bn254::Fr> {
-            a: Some(ark_bn254::Fr::rand(&mut rng)),
-            b: Some(ark_bn254::Fr::rand(&mut rng)),
+            a: Some(a),
+            b: Some(b),
             num_variables: 10,
-            num_constraints: 1 << k,
+            num_constraints: 1 << 6,
         };
-        let (pk, vk) = Groth16::<ark_bn254::Bn254>::setup(circuit, &mut rng).unwrap();
-        let c_val = circuit.a.unwrap() * circuit.b.unwrap();
+        let mut rng = ChaCha20Rng::seed_from_u64(12345);
         let proof = Groth16::<ark_bn254::Bn254>::prove(&pk, circuit, &mut rng).unwrap();
 
         // Build inputs for gadget (convert A,C to projective for wire encoding)
@@ -658,6 +595,8 @@ mod tests {
 
     #[test]
     fn test_groth16_verify_no_public_inputs_true() {
+        use crate::test_utils::DummyCircuitNoPublicInputs;
+
         // Test successful verification with empty public inputs
         let k = 6;
         let mut rng = ChaCha20Rng::seed_from_u64(99999);
@@ -691,6 +630,8 @@ mod tests {
     #[test]
     #[ignore = "Only run when modifying gadgets; this test is slow"]
     fn test_groth16_verify_no_public_inputs_false_bitflip_a() {
+        use crate::test_utils::DummyCircuitNoPublicInputs;
+
         // Test unsuccessful verification with empty public inputs - corrupt proof.a
         let k = 6;
         let mut rng = ChaCha20Rng::seed_from_u64(88888);
@@ -727,6 +668,8 @@ mod tests {
     #[test]
     #[ignore = "Only run when modifying gadgets; this test is slow"]
     fn test_groth16_verify_no_public_inputs_false_bitflip_b() {
+        use crate::test_utils::DummyCircuitNoPublicInputs;
+
         // Test unsuccessful verification with empty public inputs - corrupt proof.b
         let k = 6;
         let mut rng = ChaCha20Rng::seed_from_u64(77777);
@@ -763,6 +706,8 @@ mod tests {
     #[test]
     #[ignore = "Only run when modifying gadgets; this test is slow"]
     fn test_groth16_verify_no_public_inputs_false_bitflip_c() {
+        use crate::test_utils::DummyCircuitNoPublicInputs;
+
         // Test unsuccessful verification with empty public inputs - corrupt proof.c
         let k = 6;
         let mut rng = ChaCha20Rng::seed_from_u64(66666);
@@ -801,6 +746,8 @@ mod tests {
     fn test_groth16_verify_false_random() {
         use rand::SeedableRng;
         use rand_chacha::ChaCha20Rng;
+
+        use crate::test_utils::DummyCircuit;
 
         // Create a valid vk from a small circuit
         let k = 4;
@@ -938,6 +885,8 @@ mod tests {
 
     #[test]
     fn test_groth16_compressed_decompress_matches_proof_points() {
+        use crate::test_utils::DummyCircuit;
+
         let k = 4; // keep it small
         let mut rng = ChaCha20Rng::seed_from_u64(33333);
         let circuit = DummyCircuit::<ark_bn254::Fr> {
@@ -993,6 +942,8 @@ mod tests {
     #[test]
     #[ignore = "Only run when modifying gadgets; this test is slow"]
     fn test_groth16_verify_compressed_true_small() {
+        use crate::test_utils::DummyCircuit;
+
         let k = 4; // circuit size; pairing cost dominates anyway
         let mut rng = ChaCha20Rng::seed_from_u64(33333);
         let circuit = DummyCircuit::<ark_bn254::Fr> {
@@ -1032,6 +983,8 @@ mod tests {
         seed: u64,
         mutate: impl FnOnce(&mut Groth16VerifyInput),
     ) -> bool {
+        use crate::test_utils::DummyCircuit;
+
         let k = 4; // small circuit to keep test fast
         let mut rng = ChaCha20Rng::seed_from_u64(seed);
         let circuit = DummyCircuit::<ark_bn254::Fr> {
